requirements.txt
django
psycopg2
django-cleanup
Pillow
djangorestframework
django-suit